- delete post functionality added

+++ controllers/feed.js
@@ -143,8 +143,37 @@ exports.updatePost = (req, res, next)=>{
         })
 };
 
+exports.deletePost = (req, res, next)=>{
+
+    const postId = req.params.postId;
+    Post.findByPk(postId)
+        .then( post =>{
+            if (!post){
+                const error= new Error("Unable to find post. (error: deletepost)");
+                error.statusCode = 404;
+                throw error;
+            }
+            // check loggedIn user
+            if (post.imageUrl)
+                clearImage(post.imageUrl);
+            return post.destroy();
+        })
+            .then(result =>{
+                console.log(result);
+                res.status(200).json({message:'Post deleted!', post: result})  
+
+            })
+        .catch(err =>{
+            if (!err.statusCode){
+                err.statusCode = 500;
+            }
+            next(err);
+        })
+}
+
 
 const clearImage = filePath =>{
     filePath = path.join(__dirname, '..', filePath);
-    fs.unlink(filePath, err => console.log(err));
+    if(fs.existsSync(filePath))
+        fs.unlink(filePath, err => console.log(err))   
 }
